# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi API untuk platform OtoMeet.",
        "title": "OtoMeet API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/touring": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua daftar agenda touring komunitas dengan pagination dan search (User \u0026 Admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Touring"
                ],
                "summary": "Lihat Semua Jadwal Touring",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Nomor halaman (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah data per halaman (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cari berdasarkan nama atau tujuan touring",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan jadwal touring baru ke database (Khusus Admin Only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Touring"
                ],
                "summary": "Buat Agenda Touring Baru",
                "parameters": [
                    {
                        "description": "Data Agenda",
                        "name": "touring",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.TouringRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger403Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
        "/api/touring/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil detail satu agenda touring berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Touring"
                ],
                "summary": "Detail Jadwal Touring",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Touring ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Touring"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah data agenda touring berdasarkan ID (Khusus Admin Only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Touring"
                ],
                "summary": "Edit Jadwal Touring",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Touring ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Perubahan",
                        "name": "touring",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.TouringRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger403Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus agenda touring berdasarkan ID (Khusus Admin Only)",
                "tags": [
                    "Touring"
                ],
                "summary": "Hapus Jadwal Touring",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Touring ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger403Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.TouringRequest": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string",
                    "example": "Touring santai akhir pekan"
                },
                "harga_tiket": {
                    "type": "integer",
                    "example": 0
                },
                "image_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "kuota": {
                    "type": "integer",
                    "example": 50
                },
                "lokasi_akhir": {
                    "type": "string",
                    "example": "Lembang Park"
                },
                "lokasi_awal": {
                    "type": "string",
                    "example": "Gasibu"
                },
                "nama_touring": {
                    "type": "string",
                    "example": "Touring Lembang"
                },
                "tanggal": {
                    "type": "string",
                    "example": "2026-08-15"
                },
                "tujuan": {
                    "type": "string",
                    "example": "Lembang, Bandung"
                },
                "waktu": {
                    "type": "string",
                    "example": "08:00"
                }
            }
        },
...
        "model.Touring": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deskripsi": {
                    "type": "string"
                },
                "harga_tiket": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "kuota": {
                    "type": "integer"
                },
                "lokasi_akhir": {
                    "type": "string"
                },
                "lokasi_awal": {
                    "type": "string"
                },
                "nama_touring": {
                    "type": "string"
                },
                "tanggal": {
                    "type": "string"
                },
                "tujuan": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "waktu": {
                    "type": "string"
                }
            }
        },
...
        "utils.Swagger401Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "token tidak valid atau sudah expired"
                }
            }
        },
        "utils.Swagger403Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user tidak memiliki akses untuk fitur ini"
                }
            }
        },
        "utils.SwaggerBasicResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "string"
                },
                "error": {
                    "type": "string",
                    "example": "detail error"
                },
                "message": {
                    "type": "string",
                    "example": "detail pesan"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}